requirements.txt
# Core Web App
streamlit==1.38.0
pandas==2.2.3
plotly==5.24.1

# Backend + NLP
textblob==0.18.0
numpy==1.26.4
regex
nltk==3.9.1

# Gemini API
google-generativeai

# Optional
altair
requests
python-dotenv==1.0.1